@@ -11,9 +11,11 @@ import (
 	internalconfig "github.com/layer5io/meshery-istio/internal/config"
 	"github.com/layer5io/meshery-istio/istio/oam"
 	meshkitCfg "github.com/layer5io/meshkit/config"
+	"github.com/layer5io/meshkit/errors"
 	"github.com/layer5io/meshkit/logger"
 	"github.com/layer5io/meshkit/models"
 	"github.com/layer5io/meshkit/models/oam/core/v1alpha1"
+	"github.com/layer5io/meshkit/utils/events"
 	"gopkg.in/yaml.v2"
 )
 
@@ -23,71 +25,84 @@ type Istio struct {
 }
 
 // New initializes istio handler.
-func New(c meshkitCfg.Handler, l logger.Handler, kc meshkitCfg.Handler) adapter.Handler {
+func New(c meshkitCfg.Handler, l logger.Handler, kc meshkitCfg.Handler, ev *events.EventStreamer) adapter.Handler {
 	return &Istio{
 		Adapter: adapter.Adapter{
 			Config:            c,
 			Log:               l,
 			KubeconfigHandler: kc,
+			EventStreamer:     ev,
 		},
 	}
 }
 
 // ApplyOperation applies the operation on istio
-func (istio *Istio) ApplyOperation(ctx context.Context, opReq adapter.OperationRequest, hchan *chan interface{}) error {
+func (istio *Istio) ApplyOperation(ctx context.Context, opReq adapter.OperationRequest) error {
 	err := istio.CreateKubeconfigs(opReq.K8sConfigs)
 	if err != nil {
 		return err
 	}
 	kubeConfigs := opReq.K8sConfigs
-	istio.SetChannel(hchan)
 	operations := make(adapter.Operations)
 	err = istio.Config.GetObject(adapter.OperationsKey, &operations)
 	if err != nil {
 		return err
 	}
 
-	e := &adapter.Event{
-		Operationid: opReq.OperationID,
-		Summary:     status.Deploying,
-		Details:     "Operation is not supported",
+	e := &meshes.EventsResponse{
+		OperationId:   opReq.OperationID,
+		Summary:       status.Deploying,
+		Details:       "Operation is not supported",
+		Component:     internalconfig.ServerConfig["type"],
+		ComponentName: internalconfig.ServerConfig["name"],
 	}
-
 	switch opReq.OperationName {
 	case internalconfig.IstioOperation:
-		go func(hh *Istio, ee *adapter.Event) {
-			version := string(operations[opReq.OperationName].Versions[0])
-			stat, err := hh.installIstio(opReq.IsDeleteOperation, false, version, opReq.Namespace, "default", kubeConfigs)
-			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %s Istio service mesh", stat)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+		go func(hh *Istio, ee *meshes.EventsResponse) {
+			var err error
+			var stat, version string
+			if len(operations[opReq.OperationName].Versions) == 0 {
+				err = ErrFetchIstioVersions
+			} else {
+				version = string(operations[opReq.OperationName].Versions[len(operations[opReq.OperationName].Versions)-1])
+				stat, err = hh.installIstio(opReq.IsDeleteOperation, false, version, opReq.Namespace, "default", kubeConfigs)
+			}
+			if err != nil { //Make sure that this is a meshkit error
+				ee.Summary = fmt.Sprintf("Error while %s Istio service mesh", stat)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("Istio service mesh %s successfully", stat)
 			ee.Details = fmt.Sprintf("The Istio service mesh is now %s.", stat)
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case common.BookInfoOperation, common.HTTPBinOperation, common.ImageHubOperation, common.EmojiVotoOperation:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			appName := operations[opReq.OperationName].AdditionalProperties[common.ServiceName]
 			stat, err := hh.installSampleApp(opReq.Namespace, opReq.IsDeleteOperation, operations[opReq.OperationName].Templates, kubeConfigs)
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %s %s application", stat, appName)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while %s Istio service mesh", stat)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("%s application %s successfully", appName, stat)
 			ee.Details = fmt.Sprintf("The %s application is now %s.", appName, stat)
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case common.SmiConformanceOperation:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			name := operations[opReq.OperationName].Description
 			_, err := hh.RunSMITest(adapter.SMITestOptions{
 				Ctx:         context.TODO(),
-				OperationID: ee.Operationid,
+				OperationID: ee.OperationId,
 				Labels: map[string]string{
 					"istio-injection": "enabled",
 				},
@@ -96,60 +111,72 @@ func (istio *Istio) ApplyOperation(ctx context.Context, opReq adapter.OperationR
 				Annotations: make(map[string]string),
 			})
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %s %s test", status.Running, name)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while %s %s test", status.Running, name)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("%s test %s successfully", name, status.Completed)
 			ee.Details = ""
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case internalconfig.DenyAllPolicyOperation, internalconfig.StrictMTLSPolicyOperation, internalconfig.MutualMTLSPolicyOperation, internalconfig.DisableMTLSPolicyOperation:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			stat, err := hh.applyPolicy(opReq.Namespace, opReq.IsDeleteOperation, operations[opReq.OperationName].Templates, kubeConfigs)
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %s policy", stat)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while %s policy", stat)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("Policy %s successfully", status.Deployed)
 			ee.Details = ""
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case common.CustomOperation:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			stat, err := hh.applyCustomOperation(opReq.Namespace, opReq.CustomBody, opReq.IsDeleteOperation, kubeConfigs)
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %s custom operation", stat)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while %s custom operation", stat)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("Manifest %s successfully", status.Deployed)
 			ee.Details = ""
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case internalconfig.LabelNamespace:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			err := hh.LoadNamespaceToMesh(opReq.Namespace, opReq.IsDeleteOperation, kubeConfigs)
 			operation := "enabled"
 			if opReq.IsDeleteOperation {
 				operation = "removed"
 			}
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while labelling %s", opReq.Namespace)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while labeling %s", opReq.Namespace)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("Label updated on %s namespace", opReq.Namespace)
 			ee.Details = fmt.Sprintf("ISTIO-INJECTION label %s on %s namespace", operation, opReq.Namespace)
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case internalconfig.PrometheusAddon, internalconfig.GrafanaAddon, internalconfig.KialiAddon, internalconfig.JaegerAddon, internalconfig.ZipkinAddon:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			svcname := operations[opReq.OperationName].AdditionalProperties[common.ServiceName]
 			patches := make([]string, 0)
 			patches = append(patches, operations[opReq.OperationName].AdditionalProperties[internalconfig.ServicePatchFile])
@@ -161,26 +188,29 @@ func (istio *Istio) ApplyOperation(ctx context.Context, opReq adapter.OperationR
 			}
 
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %sing %s", operation, opReq.OperationName)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while %sing %s", operation, opReq.OperationName)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
-			ee.Summary = fmt.Sprintf("Succesfully %sed %s", operation, opReq.OperationName)
-			ee.Details = fmt.Sprintf("Succesfully %sed %s from the %s namespace", operation, opReq.OperationName, opReq.Namespace)
-			hh.StreamInfo(e)
+			ee.Summary = fmt.Sprintf("Successfully %sed %s", operation, opReq.OperationName)
+			ee.Details = fmt.Sprintf("Successfully %sed %s from the %s namespace", operation, opReq.OperationName, opReq.Namespace)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	case internalconfig.IstioVetOperation:
-		go func(hh *Istio, ee *adapter.Event) {
-			responseChan := make(chan *adapter.Event, 1)
+		go func(hh *Istio, ee *meshes.EventsResponse) {
+			responseChan := make(chan *meshes.EventsResponse, 1)
 
 			go hh.RunVet(responseChan, kubeConfigs)
 
 			for msg := range responseChan {
-				switch msg.EType {
-				case int32(meshes.EventType_ERROR):
+				switch msg.EventType {
+				case meshes.EventType_ERROR:
 					istio.StreamErr(msg, ErrIstioVet(fmt.Errorf(msg.Details)))
-				case int32(meshes.EventType_WARN):
+				case meshes.EventType_WARN:
 					istio.StreamWarn(msg, ErrIstioVet(fmt.Errorf(msg.Details)))
 				default:
 					istio.StreamInfo(msg)
@@ -190,19 +220,22 @@ func (istio *Istio) ApplyOperation(ctx context.Context, opReq adapter.OperationR
 			istio.Log.Info("Done")
 		}(istio, e)
 	case internalconfig.EnvoyFilterOperation:
-		go func(hh *Istio, ee *adapter.Event) {
+		go func(hh *Istio, ee *meshes.EventsResponse) {
 			appName := operations[opReq.OperationName].AdditionalProperties[common.ServiceName]
 			patchFile := operations[opReq.OperationName].AdditionalProperties[internalconfig.FilterPatchFile]
 			stat, err := hh.patchWithEnvoyFilter(opReq.Namespace, opReq.IsDeleteOperation, appName, operations[opReq.OperationName].Templates, patchFile, kubeConfigs)
 			if err != nil {
-				e.Summary = fmt.Sprintf("Error while %s %s application", stat, appName)
-				e.Details = err.Error()
-				hh.StreamErr(e, err)
+				ee.Summary = fmt.Sprintf("Error while %s %s application", stat, appName)
+				ee.Details = err.Error()
+				ee.ErrorCode = errors.GetCode(err)
+				ee.ProbableCause = errors.GetCause(err)
+				ee.SuggestedRemediation = errors.GetRemedy(err)
+				hh.StreamErr(ee, err)
 				return
 			}
 			ee.Summary = fmt.Sprintf("%s application %s successfully", appName, stat)
 			ee.Details = fmt.Sprintf("The %s application is now %s.", appName, stat)
-			hh.StreamInfo(e)
+			hh.StreamInfo(ee)
 		}(istio, e)
 	default:
 		istio.StreamErr(e, ErrOpInvalid)
@@ -211,7 +244,7 @@ func (istio *Istio) ApplyOperation(ctx context.Context, opReq adapter.OperationR
 	return nil
 }
 
-//CreateKubeconfigs creates and writes passed kubeconfig onto the filesystem
+// CreateKubeconfigs creates and writes passed kubeconfig onto the filesystem
 func (istio *Istio) CreateKubeconfigs(kubeconfigs []string) error {
 	var errs = make([]error, 0)
 	for _, kubeconfig := range kubeconfigs {
@@ -257,8 +290,7 @@ func (istio *Istio) CreateKubeconfigs(kubeconfigs []string) error {
 }
 
 // ProcessOAM will handles the grpc invocation for handling OAM objects
-func (istio *Istio) ProcessOAM(ctx context.Context, oamReq adapter.OAMRequest, hchan *chan interface{}) (string, error) {
-	istio.SetChannel(hchan)
+func (istio *Istio) ProcessOAM(ctx context.Context, oamReq adapter.OAMRequest) (string, error) {
 	err := istio.CreateKubeconfigs(oamReq.K8sConfigs)
 	if err != nil {
 		return "", err
